refactor: package organization (#7)
